backend/pyproject.toml
[project]
name = "fdd-backend"
version = "0.1.0"
description = "FastAPI backend for SAP solution engineering and FDD generation."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.15.1,<2.0.0",
  "docxtpl>=0.20.2,<1.0.0",
  "fastapi>=0.115.11,<1.0.0",
  "openpyxl>=3.1.5,<4.0.0",
  "psycopg[binary]>=3.2.6,<4.0.0",
  "python-docx>=1.2.0,<2.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0"
]

[dependency-groups]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-asyncio>=0.25.3,<1.0.0",
  "ruff>=0.11.0,<1.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
